Adding refactoring project with files component and containers
