+++ crud-angular/src/app/courses/components/courses-list/courses-list.component.scss
@@ -4,7 +4,7 @@
 }
 
 .action-column {
-    flex: 0 0 75px;
+    flex: 0 0 55px;
 }
 
 .action-column-header {
