+++ crud-angular/src/app/courses/courses.module.ts
@@ -4,10 +4,10 @@ import { NgModule } from '@angular/core';
 import { ReactiveFormsModule } from '@angular/forms';
 import { AppMaterialModule } from '../shared/app-material/app-material.module';
 import { SharedModule } from '../shared/shared.module';
-import { CourseFormComponent } from './course-form/course-form.component';
-import { CoursesListComponent } from './courses-list/courses-list.component';
+import { CoursesListComponent } from './components/courses-list/courses-list.component';
+import { CourseFormComponent } from './containers/course-form/course-form.component';
+import { CoursersComponent } from './containers/courses/courses.component';
 import { CoursesRoutingModule } from './courses-routing.module';
-import { CoursersComponent } from './courses/courses.component';
 
 @NgModule({
   declarations: [
